requirements.txt
aiohttp==3.6.0
beautifulsoup4==4.6.0
discord.py
urllib3
websockets==7.0
googletrans==2.4.0
lxml==4.2.1
pyowm==2.8.0
lyricsgenius==0.7.0
praw==6.0.0
dblpy
forex-python==1.1
requests
ffmpy==0.2.2
peewee==3.8.2
psycopg2-binary==2.7.7
tabulate==0.8.3
opencv-python==3.4.2.17
python-dateutil
image==1.5.27
click
timeago
psutil